+++ mep/books/admin.py
@@ -8,7 +8,7 @@ from django.utils.timezone import now
 from tabular_export.admin import export_to_csv_response
 
 from mep.accounts.admin import AUTOCOMPLETE
-from mep.books.models import Creator, CreatorType, Item
+from mep.books.models import Creator, CreatorType, Item, Subject, Format
 from mep.common.admin import CollapsibleTabularInline
 
 
@@ -30,8 +30,9 @@ class ItemCreatorInline(CollapsibleTabularInline):
 
 class ItemAdmin(admin.ModelAdmin):
     list_display = ('id', 'title', 'author_list', 'notes', 'borrow_count',
-                    'updated_at')
+                    'updated_at', 'has_uri')
     list_display_links = ('id', 'title')
+    list_filter = ('genre', 'item_format')
     inlines = [ItemCreatorInline]
     search_fields = ('mep_id', 'title', 'notes', 'creator__person__name', 'id')
     fieldsets = (
@@ -42,11 +43,18 @@ class ItemAdmin(admin.ModelAdmin):
             'fields': (
                 # ('publishers', 'pub_places'),
                 # ('volume'),
-                'uri', 'notes', 'mep_id'
+                'notes', 'mep_id'
+            )
+        }),
+        ('OCLC metadata', {
+            'fields': (
+                'uri', 'edition_uri', 'item_format', 'genre',
+                'subject_list',
+
             )
         })
     )
-    readonly_fields = ('mep_id', 'borrow_count')
+    readonly_fields = ('mep_id', 'borrow_count', 'genre', 'subject_list')
 
     actions = ['export_to_csv']
 
@@ -68,8 +76,11 @@ class ItemAdmin(admin.ModelAdmin):
     borrow_count.admin_order_field = 'event__borrow__count'
 
     #: fields to be included in CSV export
-    export_fields = ['admin_url', 'id', 'title', 'year', 'uri', 'author_list',
-                     'mep_id', 'notes']
+    export_fields = [
+        'admin_url', 'id', 'title', 'year', 'author_list', 'mep_id',
+        'uri', 'edition_uri', 'genre', 'format', 'subject_list',
+        'notes'
+    ]
 
     def csv_filename(self):
         return 'mep-items-%s.csv' % now().strftime('%Y%m%dT%H:%M:%S')
@@ -114,5 +125,18 @@ class CreatorTypeAdmin(admin.ModelAdmin):
     list_display = ('name', 'notes')
 
 
+class SubjectAdmin(admin.ModelAdmin):
+    list_display = ('name', 'uri', 'rdf_type')
+    search_fields = ('name', 'uri', 'rdf_type')
+    list_filter = ('rdf_type', )
+
+class FormatAdmin(admin.ModelAdmin):
+    list_display = ('name', 'uri')
+    # override default order to put notes last
+    fields = ('name', 'uri', 'notes')
+
+
 admin.site.register(Item, ItemAdmin)
 admin.site.register(CreatorType, CreatorTypeAdmin)
+admin.site.register(Subject, SubjectAdmin)
+admin.site.register(Format, FormatAdmin)