+++ mep/people/forms.py
@@ -97,10 +97,8 @@ class MemberSearchForm(FacetForm):
         }),
         help_text='This filter will narrow results to show only members whose \
         library records are available.')
-    # NOTE: Temporarily make submit on every click, for testing before adding
-    # to reactive search
-    sex = FacetChoiceField(label='Sex', widget=CheckboxFieldset(attrs={
-        'onchange': 'this.form.submit()'
+    sex = FacetChoiceField(label='Gender', widget=CheckboxFieldset(attrs={
+        'class': 'choice facet'
     }))
 
     def __init__(self, data=None, *args, **kwargs):